// src/api/models/AqsItemType.ts
/* eslint-disable */

export enum AqsItemType {
  Query = 'Query',
  MathAggregation = 'MathAggregation',
  StatisticsAggregation = 'StatisticsAggregation',
  Join = 'Join',
  Group = 'Group',
  And = 'And',
  Or = 'Or',
  Equals = 'Equals',
  NotEquals = 'NotEquals',
  Exists = 'Exists',
  NotExists = 'NotExists',
  StartsWith = 'StartsWith',
  WordSearch = 'WordSearch',
  GreaterThan = 'GreaterThan',
  LessThan = 'LessThan',
  GlobalAttributeSearch = 'GlobalAttributeSearch',
  GeomWithin = 'GeomWithin',
  GeomDisjoint = 'GeomDisjoint',
  GeomIntersects = 'GeomIntersects',
  ConnectedTo = 'ConnectedTo',
  NotConnectedTo = 'NotConnectedTo',
  Add = 'Add',
  Subtract = 'Subtract',
  Divide = 'Divide',
  Multiply = 'Multiply',
  Power = 'Power',
  Root = 'Root',
  Truncate = 'Truncate',
  Ceiling = 'Ceiling',
  Floor = 'Floor',
  Length = 'Length',
  Count = 'Count',
  Sum = 'Sum',
  Avg = 'Avg',
  Max = 'Max',
  Min = 'Min',
  Number = 'Number',
  Boolean = 'Boolean',
  String = 'String',
  DateTime = 'DateTime',
  Date = 'Date',
  Time = 'Time',
  Seasonal = 'Seasonal',
  RelativeDateTime = 'RelativeDateTime',
  RelativeDate = 'RelativeDate',
  ValueFromItemAttribute = 'ValueFromItemAttribute',
  Geometry = 'Geometry',
  ItemId = 'ItemId',
  AlloyId = 'AlloyId',
  Guc = 'Guc',
  Attribute = 'Attribute',
  ItemProperty = 'ItemProperty',
  Item = 'Item',
}